Debugged library; fixed issue with bitshift cmd
For Arduino Uno (ATMEGA architecture in general?), the int datatype
is 16 bit while the unsigned long datatype is 32bit. Our commands only
need 24 bits, thus were using unsigned long. However, to construct the
command, we had lines like

unsigned long cmd;
cmd = (3 << 22)

The compiler interprets "3" as an int datatype, does the bit-shift, and
finally re-casts to unsigned long. But because int is only 16bits,
most of the significant command bits were being shifted out of range.

This did not show up in Arduino Due, which was what made this bug so
mystifying. However, Arduino Due implements int as 32 bit as well.

To fix, simply do

cmd = (3UL << 22)

which will force "3" to be interpreted as an unsigned long.